Move collect{Lists,Pages,Libraries} to collection/site
