+++ src/internal/m365/collection/site/helper_test.go
@@ -4,11 +4,47 @@ import (
 	"testing"
 
 	"github.com/alcionai/clues"
+	msgraphsdk "github.com/microsoftgraph/msgraph-sdk-go"
 	"github.com/stretchr/testify/require"
 
 	"github.com/alcionai/corso/src/internal/m365/graph"
+	"github.com/alcionai/corso/src/internal/m365/support"
 	"github.com/alcionai/corso/src/pkg/account"
 )
+
+// ---------------------------------------------------------------------------
+// SharePoint Test Services
+// ---------------------------------------------------------------------------
+type MockGraphService struct{}
+
+type MockUpdater struct {
+	UpdateState func(*support.ControllerOperationStatus)
+}
+
+func (mu *MockUpdater) UpdateStatus(input *support.ControllerOperationStatus) {
+	if mu.UpdateState != nil {
+		mu.UpdateState(input)
+	}
+}
+
+//------------------------------------------------------------
+// Interface Functions: @See graph.Service
+//------------------------------------------------------------
+
+func (ms *MockGraphService) Client() *msgraphsdk.GraphServiceClient {
+	return nil
+}
+
+func (ms *MockGraphService) Adapter() *msgraphsdk.GraphRequestAdapter {
+	return nil
+}
+
+func (ms *MockGraphService) UpdateStatus(*support.ControllerOperationStatus) {
+}
+
+// ---------------------------------------------------------------------------
+// Helper functions
+// ---------------------------------------------------------------------------
 
 func createTestService(t *testing.T, credentials account.M365Config) *graph.Service {
 	adapter, err := graph.CreateAdapter(
