+++ src/internal/m365/service/sharepoint/backup_test.go
@@ -9,14 +9,9 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/alcionai/corso/src/internal/common/idname/mock"
 	"github.com/alcionai/corso/src/internal/m365/collection/drive"
-	"github.com/alcionai/corso/src/internal/m365/graph"
 	odConsts "github.com/alcionai/corso/src/internal/m365/service/onedrive/consts"
-	"github.com/alcionai/corso/src/internal/operations/inject"
 	"github.com/alcionai/corso/src/internal/tester"
-	"github.com/alcionai/corso/src/internal/tester/tconfig"
-	"github.com/alcionai/corso/src/internal/version"
 	"github.com/alcionai/corso/src/pkg/control"
 	"github.com/alcionai/corso/src/pkg/fault"
 	"github.com/alcionai/corso/src/pkg/path"
@@ -168,57 +163,4 @@ func driveRootItem(id string) models.DriveItemable {
 	item.SetFolder(models.NewFolder())
 
 	return item
-}
-
-type SharePointPagesSuite struct {
-	tester.Suite
-}
-
-func TestSharePointPagesSuite(t *testing.T) {
-	suite.Run(t, &SharePointPagesSuite{
-		Suite: tester.NewIntegrationSuite(
-			t,
-			[][]string{tconfig.M365AcctCredEnvs}),
-	})
-}
-
-func (suite *SharePointPagesSuite) SetupSuite() {
-	ctx, flush := tester.NewContext(suite.T())
-	defer flush()
-
-	graph.InitializeConcurrencyLimiter(ctx, false, 4)
-}
-
-func (suite *SharePointPagesSuite) TestCollectPages() {
-	t := suite.T()
-
-	ctx, flush := tester.NewContext(t)
-	defer flush()
-
-	var (
-		siteID = tconfig.M365SiteID(t)
-		a      = tconfig.NewM365Account(t)
-	)
-
-	creds, err := a.M365Config()
-	require.NoError(t, err, clues.ToCore(err))
-
-	ac, err := api.NewClient(creds, control.DefaultOptions())
-	require.NoError(t, err, clues.ToCore(err))
-
-	bpc := inject.BackupProducerConfig{
-		LastBackupVersion: version.NoBackup,
-		Options:           control.DefaultOptions(),
-		ProtectedResource: mock.NewProvider(siteID, siteID),
-	}
-
-	col, err := collectPages(
-		ctx,
-		bpc,
-		creds,
-		ac,
-		(&MockGraphService{}).UpdateStatus,
-		fault.New(true))
-	assert.NoError(t, err, clues.ToCore(err))
-	assert.NotEmpty(t, col)
 }
